Учет ошибки по логам

+++ config/cfg.php
@@ -40,7 +40,7 @@ return [
         App\Domain\Locale::Rus->value => ['posix' => 'ru_RU.UTF-8'],
         App\Domain\Locale::Eng->value => ['posix' => 'en_US.UTF-8'],
     ],
-    'magnet_anonymous_releaser' => env('MAGNET_ANONYMOUS_RELEASER', 3),
+    'magnet_anonymous_releaser' => (int) env('MAGNET_ANONYMOUS_RELEASER', 3),
     'sitename' => 'vacuum.kaluga',
     'sphinx' => [
         'host' => env('SPHINX_HOST', 'localhost'),
